services/query/requirements.txt
flask
python-dotenv
prometheus-client
psycopg2-binary
